BOTTEC/telegram_bot/bottest_tg/requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
inflect==7.0.0
magic-filter==1.0.12
multidict==6.0.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
sqlacodegen-v2==0.1.4
SQLAlchemy==2.0.22
typing_extensions==4.7.1
yarl==1.9.2
